cleanup some todos
