+++ labware-library/src/labware-creator/components/ConditionalLabwareRender.tsx
@@ -49,6 +49,8 @@ export const ConditionalLabwareRender = (props: Props): JSX.Element => {
       castValues = labwareFormSchema.cast(values)
       // TODO IMMEDIATELY: if we stick with this instead of single value casting, sniff this error to make sure it's
       // really a Yup validation error (see how Formik does it in `Formik.tsx`). Maybe ticket for another PR.
+      //
+      // Also, see pattern in formLevelValidation fn
     } catch (error) {}
 
     if (castValues === null) {
